requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
uvloop==0.20.0 ; sys_platform != 'win32'  # опционально ускоряет event loop в Linux
